Removed HttpServletRequest argument from MarketplaceCard.constructJSONObject
Having it as an argument didn't make a lot of sense since it mixed
controller concepts and model concepts. Perhaps a boolean could've been
used instead.

Also the extra data probably was never going to be put to use. So
including the option to get that data was unneeded complexity. It wasn't
functionality that was tested anyway.